css rework of the timesheet total calculation is a lot less brittle (sending events instead of finding the total component in the hierarchy)

+++ eHour-wicketweb/src/main/java/net/rrm/ehour/ui/timesheet/dto/ProjectTotalModel.java
@@ -22,29 +22,29 @@ public class ProjectTotalModel extends Model<Float>
 {
 	private static final long serialVersionUID = -4141505357937939279L;
 	private	TimesheetRow	row;
-	
+
 	public ProjectTotalModel(TimesheetRow row)
 	{
 		this.row = row;
 	}
-	
+
 	@Override
 	public Float getObject()
 	{
 		float	totalHours = 0;
-		
-		for (int i = 0; 
+
+		for (int i = 0;
 				i < row.getTimesheetCells().length;
 				i++)
 		{
-			if (row.getTimesheetCells()[i] != null 
-					&& row.getTimesheetCells()[i].getTimesheetEntry() != null 
+			if (row.getTimesheetCells()[i] != null
+					&& row.getTimesheetCells()[i].getTimesheetEntry() != null
 					&& row.getTimesheetCells()[i].getTimesheetEntry().getHours() != null)
 			{
-				totalHours += row.getTimesheetCells()[i].getTimesheetEntry().getHours().floatValue();
-			}			
+				totalHours += row.getTimesheetCells()[i].getTimesheetEntry().getHours();
+			}
 		}
-		
-		return Float.valueOf(totalHours);
+
+		return totalHours;
 	}
 }
